requirements.txt
art
argon2-cffi>=21.3.0
bcrypt
cryptography